Code structure changes.
